requirements.txt
asyncpg
aiogram
torch
torchvision
Pillow
numpy
nudenet
pyvips
fastapi
python-multipart
uvicorn
python-dotenv
jinja2
